refactor; add worktree id and determine main status of worktree (#301)
